Remove doom/reload
This command didn't do what was intended, and it would be too much work
to write one that looks for outdated packages and reloads them.

Instead, just restart Emacs for changes to take effect. This is only
necessary if package management is done outside of Emacs. You
(theoretically) don't have to do anything if doing it from the inside.