Fix lazy loading of elisp-mode config
elisp-mode is loaded at startup, so the usual methods won't work.
Instead, we tie a transient advice to the emacs-lisp-mode
function, *however*, this function is commonly called by various
packages to parse elisp code! So we have to make sure the emacs lisp
module only initializes the first time it is used interactively.