Defer server reloading of autoloads
Occasionally, bin/doom commands hang when generating autoloads. It
blocks endlessly after sending a message to an active Emacs server, but
never receives a proper reply.

This commit makes it less likely that this hanging will leave Doom in a
broken state (and also informs the user it is safe to abort the
process).