Rename "r" REPL prefix to "c" for connection
This leaves the "r" prefix for "refactor", and separates the common REPL
related commands into the "e" prefix for evaluating and sending forms to
the repl and "c" for managing connections with REPL sessions etc.

Also adds bindings under "cl" for sesman-link-* commands, and moves
the obsolete `cider-refresh` -> `cider-ns-refresh` under the "n" prefix.