Remove collab/{floobits,impatient-mode} modules
The state of peer programming in Emacs isn't great. The floobits module
is only one line of code and doesn't warrant its own module.
impatient-mode is a little more useful, but is too niche and not
exclusively for peer programming, so I'm not convinced it belongs in
this category. Since there are no other good options, I'm just getting
rid of the category altogether.