tools/eval: change evil-append-line to evil-insert-state
With some programs under term-mode (e.g. zsh, julia), evil-normal-state
at point-max moves to the beginning of the next blank line. When
switching to the REPL buffer, evil-append-line will append a new line,
which will cause a new prompt to be created. Don't move point and don't
call evil-append-line to fix the problem. Also, change to
evil-insert-state which is likely the most desired behavior.