Set 'evil-shift-width' to 'standard-indent' instead of 'tab-width'
'tab-width' represents the number of spaces that a <Tab> character
counts for, while 'standard-indent' represents the number of spaces
required for each level of indentation. Since 'shiftwidth' is solely
concerned with indentation, it'd make more sense to assign it to
'standard-indent'.

This also sets 'standard-indent' to 4 to retain previous behavior.