Fix DYM and add aliases
Fixed the use-package! declaration for eshell-did-you-mean since it
needed a different mode in the :after than I thought.
I also put together a quick hack to try and fix an issue with
did-you-mean, but it doesn't seem to work reliably right now. More
testing required.

Also added two aliases, one `gg` for to open magit-status and one `ff`
as an additional alias to find-file that follows the Doom keybindings
better.