Revert server-auth-dir to default
This directory is used when server-use-tcp is non-nil (which is the case
on Windows). With this changed from its default, users have to manually
specify its location when using emacsclientw.exe to connect to daemons.

This means a little more noise in `~/.emacs.d`, but Windows users have
worse things to put up with.

Fixes #3324