Remove --{emacs,doom,local}dir options
These options aren't properly supported in this version of the CLI.
Changing the localdir, for instance, doesn't affect when straight is first
bootstrapped. Chaning emacsdir doesn't matter for the first run. I'm
working on a CLI rewrite that will reimplement --doomdir and --localdir
at least, but for now it's best I just remove these.

They can still be customized using the EMACSDIR, DOOMDIR, and
DOOMLOCALDIR envvars.

Closes #3367