Fix #3650: void-variable tab-prefix-map
Seems like some OSes' package managers are serving up a build of Emacs
where it doesn't define `tab-prefix-map`, but uses it (or another
package uses it--honestly, it difficult to tell. I can't reproduce this,
but I've gotten reports of it on Arch and Ubuntu).

It'll likely go away on its own, so I'll remove this after a year or
something.