ui/doom-dashboard: have 'last-project policy fall back to CWD
Normally, the dashboard assumes the default-directory = project root of
the last buffer you had open. This is to make it easy for you to recover
from closing the last buffer in your project and wondering how to get
back. However, if the last file wasn't in a valid project, the dashboard
assumes the directory of the last buffer that did, which can be jarring.

It will now assume the last buffer's CWD if it wasn't in a valid project
root.