Update Ubuntu install guide in getting_started.org
Reason for change:
Emacs 26 is available in 19.10 and newer versions hence PPA is not
required. Emacs 27 can be installed with snap packages for now. We can
update the document when/if it is made available in official repos.

Changes:
Added apt install command for Ubuntu 19.10, 20.04 and snap command for
installing Emacs 27.