Revert company-idle-delay to package default (0.5)
Setting a default for this variable is dangerous. Set it to low and
modes with slow backends will give users the impression that this is
typing latency (they'll suspect something else, likely Doom, before
suspecting their lsp servers or company backends). Set it too high and
users may misinterpret this as latency. Turn it off (set to nil) and
I'll get a dozen "code completion isn't working" bug reports.

I've settled on company's own default of 0.5 (which is twice as slow as
Doom's original default) because responsiveness while typing is more
important (and more frustrating to deal with), but it's still enabled,
so users are more likely to notice it than assume code completion isn't
working.