Advise locate-user-emacs-file to use doom-etc-dir
This is commonly used in Emacs packages to resolve to a location in
`~/.emacs.d`, as storage or a cache. Rather than addressing each
file/directory variable as they are encountered, better to address this
at the source.

Also: this makes some file/directory settings redundant, so they were
removed.

Closes #4347