Refactor doom--help-insert-button
Changes:
- Fixes a bug where opening a file which a buffer was already visiting
didn't raise the buffer.
- The function had unused functionality where it would split a string on
'::' and then search for the text after the first '::' in the buffer;
this has been removed.
- The searching functionality has been replaced with the option to pass
a line number, which the opened buffer will jump to. This is now used by
the part of doom/help-packages that shows the places a package is configured.
- It now fails earlier. If there's an invalid file, it fails at call
time rather than when the button is pressed.
- Add a docstring