Update brew install emacs-mac instructions
When installing emacs-mac today with the command `brew install emacs-mac --with-modules` I got `Warning: Treating emacs-mac as a formula. For the cask, use railwaycat/emacsmacport/emacs-mac`. I'm not aware of any significant difference between formula vs cask, but most users are going to copy and paste the commands, see the warning, and be concerned that they messed up something. If the user has a reason to treat emacs-mac as a formula, they are probably savvy enough to use the other command.